Refactoring field variable transfers and adding array variable transferring #14391
