+++ framework/include/transfers/MultiAppNearestNodeTransfer.h
@@ -10,7 +10,7 @@
 #pragma once
 
 // MOOSE includes
-#include "MultiAppFieldTransfer.h"
+#include "MultiAppConservativeTransfer.h"
 
 // Forward declarations
 class MultiAppNearestNodeTransfer;
@@ -25,7 +25,7 @@ InputParameters validParams<MultiAppNearestNodeTransfer>();
 /**
  * Copy the value to the target domain from the nearest node in the source domain.
  */
-class MultiAppNearestNodeTransfer : public MultiAppFieldTransfer
+class MultiAppNearestNodeTransfer : public MultiAppConservativeTransfer
 {
 public:
   static InputParameters validParams();
