+++ test/tests/transfers/multiapp_copy_transfer/multivariable_copy/sub.i
@@ -1,0 +1,63 @@
+[Mesh]
+  type = GeneratedMesh
+  dim = 2
+  nx = 10
+  ny = 10
+[]
+
+[Variables]
+  [./u]
+  [../]
+  [./v]
+  [../]
+[]
+
+[Kernels]
+  [./diff_u]
+    type = Diffusion
+    variable = u
+  [../]
+  [./diff_v]
+    type = Diffusion
+    variable = v
+  [../]
+[]
+
+[BCs]
+  [./left_u]
+    type = DirichletBC
+    variable = u
+    boundary = left
+    value = 1
+  [../]
+  [./right_u]
+    type = DirichletBC
+    variable = u
+    boundary = right
+    value = 2
+  [../]
+  [./left_v]
+    type = DirichletBC
+    variable = u
+    boundary = left
+    value = 2
+  [../]
+  [./right_v]
+    type = DirichletBC
+    variable = u
+    boundary = right
+    value = 1
+  [../]
+[]
+
+[Executioner]
+  type = Transient
+  num_steps = 1
+  solve_type = 'PJFNK'
+  petsc_options_iname = '-pc_type -pc_hypre_type'
+  petsc_options_value = 'hypre boomeramg'
+[]
+
+[Outputs]
+  exodus = true
+[]
