+++ test/tests/transfers/multiapp_copy_transfer/multivariable_copy/tests
@@ -1,0 +1,12 @@
+[Tests]
+  issues = '#14391'
+  design = 'MultiAppCopyTransfer.md'
+
+  [./test]
+    type = Exodiff
+    input = master.i
+    exodiff = 'master_out.e master_out_sub0.e'
+
+    requirement = "The system shall support direct copy of multiple nonlinear variables +from+ a sub-application +to+ the main application."
+  [../]
+[]
